// apps/5r-audit/src/generated/models/_5RNewAppsModel.ts
export interface AuditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}